+++ presentation/src/main/kotlin/land/sungbin/androidprojecttemplate/ui/login/LoginScreen.kt
@@ -8,6 +8,9 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
@@ -16,6 +19,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
 import kotlinx.collections.immutable.persistentListOf
+import kotlinx.coroutines.launch
 import land.sungbin.androidprojecttemplate.R
 import team.duckie.quackquack.ui.color.QuackColor
 import team.duckie.quackquack.ui.component.QuackHeadLine2
@@ -24,6 +28,29 @@ import team.duckie.quackquack.ui.component.QuackUnderlineBody3
 
 @Composable
 fun LoginScreen(
+    viewModel: LoginViewModel,
+) {
+    val state by viewModel.state.collectAsState()
+    val coroutineScope = rememberCoroutineScope()
+
+    when (state) {
+        LoginState.Initial -> {
+            LoginContent(
+                onClickLogin = {
+                    coroutineScope.launch {
+                        viewModel.kakaoLogin()
+                    }
+                }
+            )
+        }
+
+        is LoginState.LoginFailed -> {}
+        is LoginState.KakaoLoginFailed -> {}
+    }
+}
+
+@Composable
+fun LoginContent(
     onClickLogin: () -> Unit,
 ) {
     Box(
