+++ presentation/src/main/kotlin/land/sungbin/androidprojecttemplate/ui/login/LoginViewModel.kt
@@ -36,5 +36,4 @@ class LoginViewModel @Inject constructor(
     }.onFailure {
 
     }
-
 }
